[#388] Add `conf ls` to get the configuration file locations
Introducing the command `conf ls` to `pgagroal-cli` that prints out
where the configuration files are located.
A new management key, action and three new management methods have
been added to implement the logic.
The idea is that, once the action is required, the management prints
on the socket every single `struct configuration*` path variables, and
then the management read method reads back in the same order to the
same list of values to be printed on the output screen.

In order to achieve this, two new static functions have been added to
the managemenet in order to write and read from the socket a single
path. A path is written with a size and the following path, assuming
it will never be larger than MAX_PATH or an error will be thrown.
In the case a path is NULL or empty, an empty string will be written
on the socket and so the final result is that the entry will be
printed as empty in the command output.

Documentation updated.

Shell completions updated.

Fixes also a few log_trace in the cli application.

Close #388