[#413] Fix "connection error" at the end of `status details`
This commit fix the "connection error" that appearead at the end of
the command `status details` output in pure text mode.

It also makes the numeration of connections consistent between the
JSON and text output, and changes the numeration of the connections
starting from 1 instead of 0.

It also adds some better management of the text "header bar" when
changing a section, so to avoid repetition of double lines.

Close #413