[#426] Fix a problem in `pgagroal_bind`
If the configuration `host` is set to `*`, meaning all the interfaces,
the `pgagroal_bind` function was always returning a true value even if
no available bind addresses were left.
In order to fix this, the commit tests if the `star_length` variable
has been set to something non-zero, otherwise no more sockets are
available (bound) and the function returns a `1` to indicate an error.

Close #426