[#457] Inform the user of misconfigured `log_level` debug value
If the user set the `log_level = debugX` with X being a number outside
of the valid range 1-5, the application adjusts the logging level to
the upper or lower boundary, now informing the user that her
configuration has been ignored.

Documentation updated.

Also adds `trace` as a synonim for `debug5`.

Close #457