WIP: Add new custom event loop for I/O layer
Introduce ev.h and ev.c, establishing the
foundation for the new custom event loop,
`pgagroal_ev`.

Replace previous dependencies on libev with the
custom event loop.

For Linux, implement support for io_uring with a fallback to
epoll if io_uring is unavailable.

For BSD, implement support for kqueue.

Changelog
=========

2025-02-24:
- Fix "hang" after running pgbench, caused by using both
read_message directly on sockets and io_uring. For this, I
modified `worker.c`.

2025-02-13:
- Fix ev_backend log in `configuration.c`. We cannot use
pgagroal_log_* functions during read configuration because the
log file is not set yet. Before this change, setting log type to
file leads to segmentation fault.

2025-01-18:
- delete unneeded file `bidi.c`.
- cleanup

2025-01-17:
- Fix regression added to the Debug build. `waitpid()`
inside `#if DEBUG` would block waiting for the running
child processes. Add WNOHANG to the function options so the
main process does not block.
- Remove '\n' from `pagroal_log_*()`
- Run uncrustify.sh
- Remove `printf()` debugging
- Modify CD/CI pipeline to improve function verify_running and
enable the CD/CI for kqueue, which was disabled.
- Change `pgagroal_log_error()` that informs errors in `kill()`
to `pgagroal_log_debug()`. I suppose these errors should happen
if pgagroal is not closely following up the death of its children.
Therefore, there is no need to flood the log with error messages.
Keep as debug because it is informative of something that could
be improved.

2025-01-16:
- Remove `pgagroal_ev_*_stop()` functions from main.c and vault.c
as these functions are not intended to be called from a child
process inheriting the ev loop. This does not make sense in
io_uring and for epoll and kqueue, just closing the fds is enough
as a precaution to prevent their use.
- Enhance debugging steps in `sigchld_handler()` to help debug
child processes termination.
- Add error handling to `kill()` to help debug child processes
termination.
- Allow return error on `pgagroal_ev_io_stop()` instead of exit
so the caller function can handle errors appropriately.
- Add a call to `setpgid()` when worker is created to ensure
child processes do not exit immediately if a SIGINT is issued
in the controlling terminal.
- Add a call to `pgagroal_ev_fork()` when a worker is created.

2025-01-13:
- Fix ASan report in `pgagroal_ev_io_stop()`