[#473] Fix connection status output
`pgagroal-cli` was reporting the wrong status for connections because
it was decoding the connection status via
`pgagroal_server_state_as_string` that is used to decode the status of
a server.
This commit renames the already existent function
`pgagroal_get_state_string` as `pgagroal_connection_state_as_string`
for consistency with the server
one (`pgagroal_server_state_as_string``) and makes the management
decoding using such function to get the effective status of the
connection.
Since the function `pgagroal_connection_state_as_string` keeps the
already existent statuses, the text output is not changed.
However, in the case `Unknwon` is reported, the function places also
the numeric constant in the output so that it will be easy to debug
edge cases.

Close #473