[#481] Add better error message about `Invalid master key file`
When a users file, (frontend, admins or users files) cannot be
decrypted, a generic `Invalid master key file` was raised.
This commit adds the target file that cannot be decrypted, so that the
user can understand which file has to be adjusted (e.g.,
`pgagroal-admin` usage on such file).

Makes the error messages consistent with those about a missing file.

Close #481