Fix unchecked calloc bug in network.c
At line 636 sport will always be non-NULL because it has already been checked in line 608 using early return on error.
Meanwhile result has never been checked against NULL and is subsequently used in line 696.
Probable cause of the bug is inadequate copy-and-pasting of code. I assume this changes it to the intended behavior.