ev.c: Implement io_uring infra for other IO operations and improve message handling
io_uring functions for other IO operations (send/recv) was implemented,
but is not being used because the advantage is questionable because
other communication happens serially. Therefore it is not integrated
yet. While we are at it, improve other parts of event handling for send
and recv of io_uring messages. Also fix minor bug present upstream.

Signed-off-by: Henrique de Carvalho <decarv.henrique@gmail.com>